--- server/api/match/confirm_battle.go
@@ -1,122 +1,0 @@
-package match
-
-import (
-	"context"
-	"strings"
-
-	"github.com/CryptoElementals/common/rpc/client"
-	"github.com/CryptoElementals/common/rpc/proto"
-	"github.com/CryptoElementals/common/server/api"
-	"github.com/gin-gonic/gin"
-	"github.com/go-playground/validator/v10"
-	"github.com/mitchellh/mapstructure"
-)
-
-const CONFIRM_BATTLE_LABEL = "ConfirmBattle"
-
-// ConfirmBattleRequest 请求结构体
-type ConfirmBattleRequest struct {
-	api.BaseRequest
-	GameID      uint32 `mapstructure:"GameID" validate:"required"`
-	Round       uint   `mapstructure:"Round" validate:"required"`
-	TempAddress string `mapstructure:"TempAddress" validate:"required"`
-}
-
-// ConfirmBattleResponse 响应结构体
-type ConfirmBattleResponse struct {
-	api.BaseResponse
-}
-
-type ConfirmBattleTask struct {
-	Request  *ConfirmBattleRequest
-	Response *ConfirmBattleResponse
-}
-
-// 解码请求
-func NewConfirmBattleRequest(data *map[string]interface{}) (*ConfirmBattleRequest, error) {
-	req := &ConfirmBattleRequest{}
-	err := mapstructure.Decode(*data, &req)
-	if err != nil {
-		return nil, err
-	}
-	req.BaseRequest.RequestUUID = (*data)["RequestUUID"].(string)
-	return req, nil
-}
-
-func NewConfirmBattleResponse(sessionId string) *ConfirmBattleResponse {
-	return &ConfirmBattleResponse{
-		BaseResponse: api.BaseResponse{
-			Action:      CONFIRM_BATTLE_LABEL + "Response",
-			RequestUUID: sessionId,
-		},
-	}
-}
-
-func NewConfirmBattleTask(data *map[string]interface{}) (api.Task, error) {
-	req, err := NewConfirmBattleRequest(data)
-	if err != nil {
-		return nil, err
-	}
-	task := &ConfirmBattleTask{
-		Request:  req,
-		Response: NewConfirmBattleResponse(req.BaseRequest.RequestUUID),
-	}
-
-	validate := validator.New()
-	err = validate.Struct(task.Request)
-	if err != nil {
-		return nil, err
-	}
-
-	return task, nil
-}
-
-func (task *ConfirmBattleTask) Run(c *gin.Context) (api.Response, error) {
-	// 获取玩家地址（从认证中间件设置的params中获取）
-	_params, _ := c.Get("params")
-	params, ok := _params.(*map[string]interface{})
-	if !ok {
-		task.Response.BaseResponse.RetCode = 1001
-		task.Response.BaseResponse.Message = "Failed to parse parameters"
-		return task.Response, nil
-	}
-
-	address, ok := (*params)["Address"].(string)
-	if !ok || address == "" {
-		task.Response.BaseResponse.RetCode = 1001
-		task.Response.BaseResponse.Message = "Failed to get player address"
-		return task.Response, nil
-	}
-
-	// 统一将地址转为小写
-	address = strings.ToLower(address)
-	tempAddress := strings.ToLower(task.Request.TempAddress)
-
-	// 通过gRPC调用RoomServer的ConfirmBattle
-	rpcClient := client.GetGlobalRpcClient()
-	if rpcClient == nil {
-		task.Response.BaseResponse.RetCode = 1002
-		task.Response.BaseResponse.Message = "gRPC client not initialized"
-		return task.Response, nil
-	}
-
-	req := &proto.ConfirmBattleRequest{
-		GameID:      task.Request.GameID,
-		RoundNumber: uint32(task.Request.Round),
-		PlayerAddress: &proto.PlayerAddress{
-			WalletAddress:    address,
-			TemporaryAddress: tempAddress,
-		},
-	}
-
-	_, err := rpcClient.ConfirmBattle(context.Background(), req)
-	if err != nil {
-		task.Response.BaseResponse.RetCode = 1002
-		task.Response.BaseResponse.Message = "RoomServer ConfirmBattle failed: " + err.Error()
-		return task.Response, nil
-	}
-
-	task.Response.BaseResponse.RetCode = 0
-	task.Response.BaseResponse.Message = "Successfully confirmed battle"
-	return task.Response, nil
-}
