--- server/api/user/has_collected_daily_reward.go
@@ -1,98 +1,0 @@
-package user
-
-import (
-	"strings"
-
-	"github.com/CryptoElementals/common/db"
-	"github.com/CryptoElementals/common/errors"
-	"github.com/CryptoElementals/common/log"
-	"github.com/CryptoElementals/common/server/api"
-	"github.com/gin-gonic/gin"
-	"github.com/go-playground/validator/v10"
-	"github.com/mitchellh/mapstructure"
-)
-
-const HAS_COLLECTED_DAILY_REWARD_LABEL = "HasCollectedDailyReward"
-
-type HasCollectedDailyRewardRequest struct {
-	api.BaseRequest
-}
-
-type HasCollectedDailyRewardResponse struct {
-	api.BaseResponse
-	Collected bool `json:"Collected"`
-}
-
-type HasCollectedDailyRewardTask struct {
-	Request  *HasCollectedDailyRewardRequest
-	Response *HasCollectedDailyRewardResponse
-}
-
-// 将 map 类型的数据解码为 HasCollectedDailyRewardRequest 结构体，并提取 RequestUUID
-func NewHasCollectedDailyRewardRequest(data *map[string]interface{}) (*HasCollectedDailyRewardRequest, error) {
-	req := &HasCollectedDailyRewardRequest{}
-	err := mapstructure.Decode(*data, &req)
-	if err != nil {
-		return nil, err
-	}
-	req.BaseRequest.RequestUUID = (*data)["RequestUUID"].(string)
-	return req, nil
-}
-
-func NewHasCollectedDailyRewardResponse(sessionId string) *HasCollectedDailyRewardResponse {
-	return &HasCollectedDailyRewardResponse{
-		BaseResponse: api.BaseResponse{
-			Action:      HAS_COLLECTED_DAILY_REWARD_LABEL + "Response",
-			RequestUUID: sessionId,
-		},
-	}
-}
-
-func NewHasCollectedDailyRewardTask(data *map[string]interface{}) (api.Task, error) {
-	req, err := NewHasCollectedDailyRewardRequest(data)
-	if err != nil {
-		return nil, err
-	}
-	task := &HasCollectedDailyRewardTask{
-		Request:  req,
-		Response: NewHasCollectedDailyRewardResponse(req.BaseRequest.RequestUUID),
-	}
-
-	validate := validator.New()
-	err = validate.Struct(task.Request)
-	if err != nil {
-		return nil, err
-	}
-
-	return task, nil
-}
-
-func (task *HasCollectedDailyRewardTask) Run(c *gin.Context) (api.Response, error) {
-	// 从请求参数中获取用户地址（由中间件设置）
-	_params, _ := c.Get("params")
-	params, ok := _params.(*map[string]interface{})
-	if !ok {
-		log.Errorf("%s, params assert failed", task.Request.RequestUUID)
-		return nil, errors.MissingLoginCookie()
-	}
-
-	address, ok := (*params)["Address"].(string)
-	if !ok || address == "" {
-		log.Errorf("%s, no address found in params", task.Request.RequestUUID)
-		return nil, errors.MissingLoginCookie()
-	}
-
-	// 将地址转换为小写，确保与数据库中存储的格式一致
-	lowercaseAddress := strings.ToLower(address)
-
-	// 检查用户是否已领取今日奖励
-	collected, err := db.HasCollectedDailyReward(lowercaseAddress)
-	if err != nil {
-		log.Errorf("%s, failed to check daily reward collection for address %s: %v", task.Request.RequestUUID, lowercaseAddress, err)
-		return nil, errors.GetUserProfileFailed(lowercaseAddress)
-	}
-
-	task.Response.Collected = collected
-	log.Infof("%s, daily reward collection status checked for address %s: %v", task.Request.RequestUUID, lowercaseAddress, collected)
-	return task.Response, nil
-}
